feat: A big overhaul
This commit sees the revamp of the CLI, setting it up for
expansion of features potentially in the future.

We also find the root cause (and fixed) the issue with the quotes being
incorrect on the JSON -> Snippet mappings.

Now the goal is to test this tool thoroughly before we make the next
release (Expected date is hopefully October 1, 2022).