feat: Add optional description support
This change probably offers the greatest move toward flexibility as
it allows for an optional `description` field to be on the
source snippet.

If it doesn't exist, then we pass a `None` and skip the serialization
of that `None` field in the final output.

We also add several unit tests to ensure output matches expectations and
the Optional typing doesn't mess with existing functionality.